requirements.txt
Django==3.2.19
mysqlclient==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
python-dotenv==1.0.0
django-filter==22.1